Centralize PackageIdentityArgument
