+++ src/Cli/dotnet/Commands/Tool/ToolCommandParser.cs
@@ -40,4 +40,5 @@ internal static class ToolCommandParser
 
         return command;
     }
+
 }
