+++ src/Cli/dotnet/Commands/Tool/Update/ToolUpdateCommandParser.cs
@@ -9,7 +9,7 @@ namespace Microsoft.DotNet.Cli.Commands.Tool.Update;
 
 internal static class ToolUpdateCommandParser
 {
-    public static readonly CliArgument<(string PackageId, string Version)> PackageIdArgument = ToolInstallCommandParser.PackageIdArgument;
+    public static readonly CliArgument<(string PackageId, string Version)> PackageIdArgument = CommonArguments.PackageIdentityArgument;
 
     public static readonly CliOption<bool> UpdateAllOption = ToolAppliedOption.UpdateAllOption;
